requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
flask>=3.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.2
langchain-groq>=0.1.0

# Vector DB and Embeddings
chromadb>=0.4.22

# PDF Processing
pdfplumber>=0.10.0

# HTTP Client
requests>=2.31.0
